Trying passing func delegate
