+++ RightMoveApp/UserControls/SearchParamsViewModel.cs
@@ -1,7 +1,12 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
 using RightMove.DataTypes;
+using static RightMove.Desktop.UserControls.AutoCompleteComboBox;
 
 namespace RightMove.Desktop.UserControls
 {
@@ -114,6 +119,7 @@ namespace RightMove.Desktop.UserControls
 		private PropertyTypeEnum _propertyType;
         private ObservableCollection<string> _regionStrings;
         private RightMoveRegion _selectedRightMoveRegion;
+        private AutocompleteSearchCallback _rightMoveFunc = DefaultFunc;
 
         public PropertyTypeEnum PropertyType
 		{
@@ -146,6 +152,27 @@ namespace RightMove.Desktop.UserControls
         {
             get => _regionStrings;
             set => SetProperty(ref _regionStrings, value);
+        }
+
+        public static AutocompleteSearchCallback DefaultFunc = async (text, token) =>
+        {
+            var regionService = new RightMoveRegionService();
+
+            try
+            {
+                var items = (await regionService.SearchAsync(text, token)).ToList();
+                return items;
+            }
+            catch (TaskCanceledException)
+            {
+                return new List<RightMoveRegion>();
+            }
+        };
+
+        public AutocompleteSearchCallback RightMoveFunc
+        {
+            get => _rightMoveFunc;
+            set => SetProperty(ref _rightMoveFunc, value);
         }
