+++ RightMoveApp/View/SearchParamsView.xaml
@@ -35,7 +35,8 @@
         <userControls:AutoCompleteComboBox Grid.Row="0" Grid.Column="1"
                                            ItemsSource="{Binding ElementName=uc, Path=SearchString}" 
                                            SelectedRightMoveRegion="{Binding Path=SelectedRightMoveRegion, Mode=TwoWay}"
-                                           MinimumCharactersForSearch="2"/>
+                                           MinimumCharactersForSearch="2"
+                                           AutocompleteSearchFunc="{Binding RightMoveFunc}"/>
 
         <Label Grid.Column="0" Grid.Row="1" Content="Search radius"/>
         <ComboBox Template="{DynamicResource ComboBoxTemplate1}" Grid.Column="1" Grid.Row="1" Name="comboSearchRadius"
